ajustes cambios nombres en retroalimentacion

+++ microservicio/aplicacion/src/main/java/com/ceiba/scotter/consulta/ManejadorBuscarPrecioScotter.java
@@ -15,7 +15,7 @@ public class ManejadorBuscarPrecioScotter {
         this.daoScotter = daoScotter;
     }
 
-    public List<DtoScotter> ejecutar(Integer precio) {
+    public List<DtoScotter> ejecutar(double precio) {
         return this.daoScotter.buscarPrecio(precio);
     }
 }
