+++ microservicio/dominio/src/test/java/com/ceiba/scotter/servicio/testdatabuilder/ScotterTestDataBuilder.java
@@ -2,8 +2,6 @@ package com.ceiba.scotter.servicio.testdatabuilder;
 
 import com.ceiba.scotter.modelo.entidad.Scotter;
 
-
-
 public class ScotterTestDataBuilder {
     private Long id;
     private String marca;
@@ -12,11 +10,11 @@ public class ScotterTestDataBuilder {
     private String ciudad;
     private Long vendedor;
     private String estado;
-    private  String foto;
+    private String foto;
 
     public ScotterTestDataBuilder(String marca, String modelo, double precio, String ciudad,
                                   Long vendedor, String foto) {
-         this.marca = marca;
+        this.marca = marca;
         this.modelo = modelo;
         this.precio = precio;
         this.ciudad = ciudad;
@@ -25,7 +23,7 @@ public class ScotterTestDataBuilder {
     }
 
     public ScotterTestDataBuilder() {
-        id= 1L;
+        id = 1L;
         marca = "toyota";
         modelo = "2022";
         precio = 6000000.0;
@@ -33,7 +31,6 @@ public class ScotterTestDataBuilder {
         vendedor = 2L;
         estado = "disponible";
         foto = "https://www.google.com/search?q=foto+hd+scooter&rlz=1C1UUXU_esCO947CO947&source=lnms&tbm=isch&sa=X&ved=2ahUKEwjcyNSkxPTxAhXkRjABHb2SBecQ_AUoAXoECAEQAw&biw=1600&bih=757";
-
     }
 
     public ScotterTestDataBuilder conId(Long id) {
@@ -46,12 +43,12 @@ public class ScotterTestDataBuilder {
         return this;
     }
 
-    public ScotterTestDataBuilder conPrecio(int precio) {
+    public ScotterTestDataBuilder conPrecio(double precio) {
         this.precio = precio;
         return this;
     }
 
     public Scotter build() {
-        return new Scotter(id,marca, modelo,precio,ciudad,vendedor,foto);
+        return new Scotter(id, marca, modelo, precio, ciudad, vendedor, foto);
     }
 }
